Expose the output object file for each C++ source file in the C++ tooling model.
