+++ subprojects/tooling-builders-native/src/main/java/org/gradle/language/cpp/internal/tooling/DefaultCppProjectModel.java
@@ -1,0 +1,45 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.language.cpp.internal.tooling;
+
+import org.gradle.tooling.internal.gradle.DefaultProjectIdentifier;
+
+import java.io.Serializable;
+
+public class DefaultCppProjectModel implements Serializable {
+    private final DefaultProjectIdentifier projectIdentifier;
+    private final DefaultCppComponentModel mainComponent;
+    private final DefaultCppComponentModel testComponent;
+
+    public DefaultCppProjectModel(DefaultProjectIdentifier projectIdentifier, DefaultCppComponentModel mainComponent, DefaultCppComponentModel testComponent) {
+        this.projectIdentifier = projectIdentifier;
+        this.mainComponent = mainComponent;
+        this.testComponent = testComponent;
+    }
+
+    public DefaultProjectIdentifier getProjectIdentifier() {
+        return projectIdentifier;
+    }
+
+    public DefaultCppComponentModel getMainComponent() {
+        return mainComponent;
+    }
+
+    public DefaultCppComponentModel getTestComponent() {
+        return testComponent;
+    }
+}
