+++ subprojects/tooling-builders-native/src/main/java/org/gradle/language/cpp/internal/tooling/DefaultCppSharedLibraryModel.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.language.cpp.internal.tooling;
+
+import org.gradle.tooling.internal.protocol.cpp.InternalCppSharedLibrary;
+
+public class DefaultCppSharedLibraryModel extends DefaultCppBinaryModel implements InternalCppSharedLibrary {
+    public DefaultCppSharedLibraryModel(String name, String baseName, DefaultCompilationDetails compilationDetails, DefaultLinkageDetails linkageDetails) {
+        super(name, baseName, compilationDetails, linkageDetails);
+    }
+}
