+++ subprojects/tooling-builders-native/src/main/java/org/gradle/language/cpp/internal/tooling/DefaultSourceFile.java
@@ -1,0 +1,38 @@
+/*
+ * Copyright 2018 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.gradle.language.cpp.internal.tooling;
+
+import java.io.File;
+import java.io.Serializable;
+
+public class DefaultSourceFile implements Serializable {
+    private final File sourceFile;
+    private final File objectFile;
+
+    public DefaultSourceFile(File sourceFile, File objectFile) {
+        this.sourceFile = sourceFile;
+        this.objectFile = objectFile;
+    }
+
+    public File getSourceFile() {
+        return sourceFile;
+    }
+
+    public File getObjectFile() {
+        return objectFile;
+    }
+}
